refactor(stable): generate extractors types from OpenAPI
Switch extractors model types to codegen and keep manual types only for
SDK-specific query and schema helpers.
