+++ packages/stable/src/api/extractors/codegen.json
@@ -1,0 +1,20 @@
+{
+  "service": "extractors",
+  "filter": {
+    "serviceName": "extractors",
+    "relevantReferenceNames": [
+      "Extractor",
+      "Release",
+      "SourceSystem",
+      "Solution",
+      "ReleaseId",
+      "ExtractorId",
+      "Artifact",
+      "Link",
+      "Changelog"
+    ]
+  },
+  "inlinedSchemas": {
+    "autoNameRequest": true
+  }
+}
