+++ packages/stable/src/api/extractors/types.gen.ts
@@ -1,0 +1,232 @@
+// Do not modify this file!
+// It was generated by the command "yarn codegen".
+// Instead update the code generation logic or the OpenAPI document.
+
+/**
+ * An artifact for an extractor release.
+ */
+export interface Artifact {
+  /**
+   * Display Name
+   * Human readable name of the artifact.
+   */
+  displayName?: string;
+  /**
+   * Link
+   * Link to obtain a temporary download link for the artifact.
+   */
+  link: string;
+  /**
+   * Name
+   * Filename of the artifact.
+   */
+  name: string;
+  /**
+   * Platform
+   * Platform the extractor runs on. One of "windows", "linux", "macos", "docs", or "all".
+   */
+  platform: 'windows' | 'linux' | 'macos' | 'docs' | 'all';
+}
+/**
+ * Extractor changelog, uses the [keep-a-changelog](https://keepachangelog.com/en/1.1.0/) format
+ */
+export interface Changelog {
+  /**
+   * Added
+   * Features added to this release.
+   */
+  added?: string[];
+  /**
+   * Changed
+   * Changes made to the extractor in this release.
+   */
+  changed?: string[];
+  /**
+   * Deprecated
+   * Features that have been deprecated, but not yet removed.
+   */
+  deprecated?: string[];
+  /**
+   * Fixed
+   * Bugs fixed in this release.
+   */
+  fixed?: string[];
+  /**
+   * Removed
+   * Features that have been removed.
+   */
+  removed?: string[];
+  /**
+   * Security
+   * Changes that affect security.
+   */
+  security?: string[];
+}
+/**
+ * The external ID provided by the client. Must be unique for the resource type.
+ * @example my.known.id
+ */
+export type CogniteExternalId = string;
+/**
+ * The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
+ * @format int64
+ * @min 0
+ * @example 1730204346000
+ */
+export type EpochTimestamp = number;
+/**
+ * An extractor instance
+ */
+export interface Extractor {
+  /**
+   * Description
+   * Short description of the extractor.
+   */
+  description: string;
+  /**
+   * Documentation
+   * Longer markdown documentation.
+   */
+  documentation?: string;
+  /** The external ID provided by the client. Must be unique for the resource type. */
+  externalId: CogniteExternalId;
+  /**
+   * Image Url
+   * Image URL for extractor logo.
+   */
+  imageUrl?: string;
+  /**
+   * Latest Version
+   * The last released version of the extractor.
+   */
+  latestVersion?: string;
+  /**
+   * Links
+   * A list of external links.
+   */
+  links?: Link[];
+  /**
+   * Name
+   * Name of this extractor.
+   */
+  name: string;
+  /**
+   * Tags
+   * A list of tags, used for searching and filtering.
+   */
+  tags?: string[];
+  /**
+   * Type
+   * Type of extractor. "Global" means that it is globally available and maintained by cognite. "Community" means that it is unofficial and community maintained. "Unreleased" means that it is not yet publicly available and it is visible only as a prerelease or private beta. "Hosted" means it is a hosted extractor that runs as part of the CDF platform.
+   */
+  type: 'global' | 'community' | 'unreleased' | 'hosted';
+}
+/**
+ * Identifier for an extractor
+ */
+export interface ExtractorId {
+  /** The external ID provided by the client. Must be unique for the resource type. */
+  externalId: CogniteExternalId;
+}
+/**
+ * Status of this item. "Global" means that it is maintained by Cognite and considered official. "Community" means that it is community made and maintained. "Unreleased" means that it is not yet released and it is visible only as a preview or closed beta.
+ */
+export type ItemType = 'global' | 'community' | 'unreleased';
+/**
+ * Link to an external resource for an extractor
+ */
+export interface Link {
+  /**
+   * Name
+   * Human readable link name.
+   */
+  name: string;
+  /** Enumeration of link types. */
+  type: LinkType;
+  /**
+   * Url
+   * Link URL.
+   */
+  url: string;
+}
+/**
+ * Enumeration of link types.
+ */
+export type LinkType = 'generic' | 'externalDocumentation';
+/**
+ * A release of an extractor
+ */
+export interface Release {
+  /**
+   * Artifacts
+   * List of artifacts included in this release.
+   */
+  artifacts: Artifact[];
+  /** Extractor changelog, uses the [keep-a-changelog](https://keepachangelog.com/en/1.1.0/) format */
+  changelog?: Changelog;
+  /** The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds. */
+  createdTime?: EpochTimestamp;
+  /**
+   * Description
+   * Short description of this release. Details are found in the changelog.
+   */
+  description?: string;
+  /** The external ID provided by the client. Must be unique for the resource type. */
+  externalId: CogniteExternalId;
+  /**
+   * Version
+   * Release version number, uses semantic versioning.
+   */
+  version: string;
+}
+/**
+ * Identifier for a release
+ */
+export interface ReleaseId {
+  /** The external ID provided by the client. Must be unique for the resource type. */
+  externalId: CogniteExternalId;
+  /**
+   * Version
+   * Release version number, uses semantic versioning.
+   */
+  version: string;
+}
+/**
+ * A solution represents a connection between an extractor and a source system it can be configured to read from.
+ */
+export interface Solution {
+  /** Long-form documentation of the solution, describing how to use the referenced extractor to connect to the referenced source. */
+  documentation?: string;
+  /** The external ID provided by the client. Must be unique for the resource type. */
+  externalId: CogniteExternalId;
+  /** External ID of the extractor of this solution. */
+  extractorExternalId?: string;
+  /** Solution name. */
+  name: string;
+  /** External ID of the source system of this solution. */
+  sourceSystemExternalId: string;
+  /** Status of this item. "Global" means that it is maintained by Cognite and considered official. "Community" means that it is community made and maintained. "Unreleased" means that it is not yet released and it is visible only as a preview or closed beta. */
+  type?: ItemType;
+}
+/**
+ * A source system representing a source extractors may read from.
+ */
+export interface SourceSystem {
+  /** Short description of the source system. */
+  description: string;
+  /** Long-form description of the source system. */
+  documentation?: string;
+  /** The external ID provided by the client. Must be unique for the resource type. */
+  externalId: CogniteExternalId;
+  /** URL of a publicly hosted logo for the source system. */
+  imageUrl?: string;
+  /** Source system name. */
+  name: string;
+  /**
+   * Tags
+   * A list of tags, used for searching and filtering.
+   */
+  tags?: string[];
+  /** Status of this item. "Global" means that it is maintained by Cognite and considered official. "Community" means that it is community made and maintained. "Unreleased" means that it is not yet released and it is visible only as a preview or closed beta. */
+  type: ItemType;
+}
