+++ packages/stable/src/exports.gen.ts
@@ -1,5 +1,6 @@
 // Copyright 2022 Cognite AS
 export type {
+  CogniteExternalId,
   EpochTimestamp,
   InstanceExternalId,
 } from './types.gen';
@@ -40,7 +41,6 @@ export type {
   AnnotationsView,
 } from './api/annotations/types.gen';
 export type {
-  CogniteExternalId,
   CogniteInstanceId,
   CogniteInternalId,
   Document,
@@ -133,6 +133,19 @@ export type {
   PassageSourceFile,
   Space,
 } from './api/documents/types.gen';
+export type {
+  Artifact,
+  Changelog,
+  Extractor,
+  ExtractorId,
+  ItemType,
+  Link,
+  LinkType,
+  Release,
+  ReleaseId,
+  Solution,
+  SourceSystem,
+} from './api/extractors/types.gen';
 export type {
   CursorQueryParameter,
   IncludeGlobalQueryParameter,
